[Player] use dbc values for resource base/regen/active/haste/start_at

+++ engine/dbc/spell_data.cpp
@@ -48,7 +48,8 @@ const power_type_data_t& power_type_data_t::find( power_e type, bool ptr )
   auto raw_type = static_cast<int>( type );
   const auto __data = data( ptr );
 
-  assert( raw_type >= 0 && raw_type < __data.size() && "power_type_data_t: power_e type index out of bounds" );
+  assert( raw_type >= 0 && raw_type < static_cast<int>( __data.size() ) &&
+          "power_type_data_t: power_e type index out of bounds" );
 
   return __data[ raw_type ];
 }
